Refactor: MyArticlePreview 로직부분 훅으로 분리
