+++ src/components/pages/Mypage/hooks/index.js
@@ -1,5 +1,6 @@
 import useMypageProfile from './useMypageProfile';
 import useProfileSection from './useProfileSection';
+import useMyArticlePreview from './useMyArticlePreview';
 import useMyArticleBoard from './useMyArticleBoard';
 
-export { useMypageProfile, useProfileSection, useMyArticleBoard };
+export { useMypageProfile, useProfileSection, useMyArticlePreview, useMyArticleBoard };
